Adding "logcall", similar to "syscall", but unlike syscalls, the call made triggers how log should behave.
For example, in this implementation we have "Auth.Login", where any log writer process that can listen to this will be declared and then change the key-values to fetch them from the variables given and respond to which event means an "Auth.Login"

As of now, it can be only used with ccraft. It would make more sense to move to pcap-ng ASAP so we can also have a pcap support for this, instead of stretching pcap in the way we did it already for endpoint data.